+++ Sources/Dependencies/Internal/Deprecations.swift
@@ -1,3 +1,51 @@
+// MARK: - Deprecated after 1.6.3
+
+#if canImport(SwiftUI) && compiler(>=6)
+  import SwiftUI
+
+  @available(iOS 18, macOS 15, tvOS 18, watchOS 11, visionOS 2, *)
+  extension PreviewTrait where T == Preview.ViewTraits {
+    @available(
+      *, deprecated, message: """
+        Use 'withDependencies' or 'prepareDependencies' from the body of the preview, instead.
+        """
+    )
+    public static func dependency<Value>(
+      _ keyPath: WritableKeyPath<DependencyValues, Value> & Sendable,
+      _ value: Value
+    ) -> PreviewTrait {
+      .dependencies { $0[keyPath: keyPath] = value }
+    }
+
+    @available(
+      *, deprecated, message: """
+        Use 'withDependencies' or 'prepareDependencies' from the body of the preview, instead.
+        """
+    )
+    public static func dependency<Value: TestDependencyKey>(
+      _ value: Value
+    ) -> PreviewTrait where Value == Value.Value {
+      .dependencies { $0[Value.self] = value }
+    }
+
+    @available(
+      *, deprecated, message: """
+        Use 'withDependencies' or 'prepareDependencies' from the body of the preview, instead.
+        """
+    )
+    public static func dependencies(
+      _ updateValuesForPreview: (inout DependencyValues) -> Void
+    ) -> PreviewTrait {
+      var copy = previewValues
+      defer { previewValues = copy }
+      updateValuesForPreview(&copy)
+      return PreviewTrait()
+    }
+  }
+
+  nonisolated(unsafe) var previewValues = DependencyValues(context: .preview)
+#endif
+
 // MARK: - Deprecated after 0.4.2
 
 extension AsyncStream {
